chore: remove string constants and name utils from formula-types.ts [PT-186452591]
