+++ v3/src/models/formula/functions/lookup-functions.ts
@@ -1,6 +1,7 @@
 import { ConstantNode, MathNode } from "mathjs"
 import { FormulaMathJsScope } from "../formula-mathjs-scope"
-import { DisplayNameMap, ILookupDependency, isConstantStringNode, rmCanonicalPrefix } from "../formula-types"
+import { DisplayNameMap, ILookupDependency, isConstantStringNode } from "../formula-types"
+import { rmCanonicalPrefix } from "../utils/name-mapping-utils"
 import { UNDEF_RESULT, equal, evaluateNode } from "./function-utils"
 import type { IDataSet } from "../../data/data-set"
 import t from "../../../utilities/translation/translate"
